find S algorithm
It worked on the positive hypothesis according to these rules
Initialize h to the most specific hypothesis in H
For each positive training instance x
For each attribute constraint a i in h :
If the constraint a i in h is satisfied by x Then do nothing
Else replace a i in h by the next more general constraint that is satisfied by x
Output hypothesis h
Guaranteed to output the most specific hypothesis within H that is consistent with the positive training examples.
Notice that negative examples are ignored.